sistema/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Brotli==1.1.0
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-extensions==0.7.1
drf-spectacular==0.26.5
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
packaging==23.2
Pillow==10.0.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyMySQL==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.4
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
whitenoise==6.6.0
